Refactor: Controllers folders
Organized all the controllers in folders that compares with use-cases to make the analise easier.
